+++ src/pages/login/LoginPage.jsx
@@ -1,34 +1,38 @@
 import './loginPage.scss';
 import '../../reset.scss';
+import { useEffect } from 'react';
 import { PageContainer } from '../../components';
 import LoginInputBox from './login-input/Email.Pwd.Input.component';
 import SnsButtons from './sns/LoginSnsnButtons';
-import { logout } from './loginUtil';
-import { useRegisterStore } from '../register/RegisterStore';
-import { useEffect } from 'react';
+import { useRegisterStore } from '../../states/register/RegisterStore';
+import { logout } from '../../services/login/loginApi';
+import { loginStateStore } from '../../states/login/loginStore';
 
 export default function Login() {
-  const { reset } = useRegisterStore();
+	const { reset } = useRegisterStore();
+	const { loginInfo, loginedStateStorereset } = loginStateStore();
+	useEffect(() => {
+		reset();
+		if (loginInfo) {
+			loginedStateStorereset();
+		}
+	}, []);
 
-  useEffect(() => {
-    reset();
-  }, []);
+	const Test = async () => {
+		const result = await logout();
+		console.log(result);
+	};
 
-  const Test = async () => {
-    const result = await logout();
-    console.log(result);
-  };
-
-  return (
-    <PageContainer className={'login-container'}>
-      <div className='login-wrap'>
-        <div className='login-page up'>
-          <h2>로그인</h2>
-        </div>
-        <LoginInputBox />
-        <SnsButtons />
-        <button onClick={Test}>로그아웃</button>
-      </div>
-    </PageContainer>
-  );
+	return (
+		<PageContainer className={'login-container'}>
+			<div className='login-wrap'>
+				<div className='login-page up'>
+					<h2>로그인</h2>
+				</div>
+				<LoginInputBox />
+				<SnsButtons />
+				<button onClick={Test}>로그아웃</button>
+			</div>
+		</PageContainer>
+	);
 }
