+++ src/pages/login/loginInterCepter/LoginInterceptor.jsx
@@ -1,14 +1,20 @@
-import { Navigate, useLocation } from "react-router-dom";
+import { Navigate, useLocation } from 'react-router-dom';
 
 const LoginInterceptor = ({ children }) => {
-  const isAuthenticated = !!sessionStorage.getItem("Logined"); // 로그인 여부 확인
-  const location = useLocation();
+	const isAuthenticated = !!sessionStorage.getItem('Logined'); // 로그인 여부 확인
+	const location = useLocation();
 
-  if (!isAuthenticated) {
-    return <Navigate to="/login" state={{ from: location }} replace />;
-  }
+	if (!isAuthenticated) {
+		return (
+			<Navigate
+				to='/login'
+				state={{ from: location }}
+				replace
+			/>
+		);
+	}
 
-  return children;
+	return children;
 };
 
 export default LoginInterceptor;
